Changes to log to kinesis streams
